backend/requirements.txt
fastapi
uvicorn
python-dotenv
azure-identity
azure-ai-inference
